Remove IRC interface
Whilst an IRC interface might be fun, it's not something that we
actually need.  At the competition, all administration of the system
is performed in a terminal on the comp VM or on the same network.  We
don't need this second interface to the system.

This patch removes the IRC interface, removing 56 lines of code that
we no longer need to maintain.