Rip out the twisted reactor, replace with plain redis calls.
This moves this source away from being a server that needs to be
running in the background to being a library of useful tools to
interact with a redis database.

For the moment, the resulting code is pretty dead as there's now
very little which connects the pieces together. These will appear
shortly.

Change-Id: Ib7db4f8f399684385d4c23acc68e8d217f0b5f04