Add a schema for the redis datastore, update doc comments.
Add a schema to provide a handy reference for developing against.
While it's not expected that this will be usedby external clients
(they should use the API which poltergeist presents), it's very
handy to have this while developing poltergeist.

Update several outdated doc comments which creating this schema
showed up.

Change-Id: Ic6feb3bb6ef31cf152a4a48b995671edb0336562