Separate out commands from database wrappers.
This provides a clearer separation between the portions of the code
which wrap access to redis and those which pump that data outwards.
Whatever outwards is going to mean this year.

This also separates the tests, and exposes 4 more tests which had
name clashes and were thus hiding each other.

Change-Id: I264e79338039b434728f14fa1c47f073550742ed