Change to using a ConfigParser based config.ini.
This makes it simpler to add local options, without needing to copy
the entire yaml file. Also remove most of the irrelevant options
during the conversion.
Make the redis_connection use this new mechanism, and have it follow
a similar pattern so that the connection is immediately available.

Remove config imports from display files since this is no longer needed.

Change-Id: I236b761bcccfa91b9e6c678bc6f0653db317049a